refactor: TeamChangeModal 메인 모달 리팩토링 완료
- 기존 260줄 TeamChangeModal.tsx를 70줄 index.tsx로 간소화
- TimelineSection, VotingSection 컴포넌트 조합
- 기존 파일 삭제 및 디렉토리 구조 완성
- 모든 로직을 하위 컴포넌트로 분리하여 가독성 및 유지보수성 향상
